requirements.txt
gradio
python-dotenv
openai

# image captioning
transformers
torch